+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.41](https://github.com/mbc-net/mbc-cqrs-serverless-web/compare/v0.0.40...v0.0.41) (2026-02-10)
+
+### Features
+
+- update RichTextEditor and field-editor with enhanced toolbar options and color palette ([586bd6f](https://github.com/mbc-net/mbc-cqrs-serverless-web/commit/586bd6fadd0d9291bb2952c8a54b6457a42570ad))
+
+## [0.0.40](https://github.com/mbc-net/mbc-cqrs-serverless-web/compare/v0.0.39...v0.0.40) (2026-01-27)
+
+### Features
+
+- enhance ReactQuill integration with custom block registration ([eece858](https://github.com/mbc-net/mbc-cqrs-serverless-web/commit/eece858a8b3e98d64b70af13abcdba53ca47b343))
+
 ## [0.0.39](https://github.com/mbc-net/mbc-cqrs-serverless-web/compare/v0.0.38...v0.0.39) (2025-12-18)
 
 ### Bug Fixes